Add a minimalistic band annotation pipeline in spaCy
This pipeline introduces spaCy as a dependency and create a
minimalistic band annotation pipeline on top of it. The pipeline
tokenize the text it gets and use a gazetteer to find possible band
names.

Next steps are to look at better unit tests environment for my
gazetteer stored in a database and to start collecting ground truth
data that will help us iterate on the matching logic.